basically working renderer, some stuf still needs to be done but I need to wait for Correntin at this point

+++ src/dist_functions.c
@@ -6,7 +6,7 @@
 /*   By: awoimbee <awoimbee@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/08 22:02:07 by awoimbee          #+#    #+#             */
-/*   Updated: 2019/01/12 13:32:56 by awoimbee         ###   ########.fr       */
+/*   Updated: 2019/01/12 18:24:41 by awoimbee         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,9 +37,10 @@ float		dist_plane(const union u_object *obj, const t_ray ray)
 	float	dist;
 
 	denom = vec3_dot(obj->plane.norm, ray.dir);
-	if (fabsf(denom) > 1e-6)
+	if (fabs(denom) > 1e-6)
 	{
-		dist = vec3_dot(vec3_sub(obj->plane.orig, ray.org), obj->plane.norm) / denom;
+		dist = 
+		vec3_dot(vec3_sub(obj->plane.orig, ray.org), obj->plane.norm) / denom;
 		if (dist > 0)
 			return (dist);
 	}
@@ -54,7 +55,8 @@ float		dist_disk(const union u_object *obj, const t_ray ray)
 
 	if ((dist = dist_plane(obj, ray)) != -1)
 	{
-		v = vec3_sub(vec3_add(ray.org, vec3_multf(ray.dir, dist)), obj->disk.orig);
+		v =
+		vec3_sub(vec3_add(ray.org, vec3_multf(ray.dir, dist)), obj->disk.orig);
 		disk2 = vec3_dot(v, v);
 		if (disk2 <= obj->disk.radius2)
 			return (dist);
